Move global intervals to fleet scheduler
