+++ src/index.ts
@@ -1,16 +1,26 @@
 import { isMaster } from "cluster";
 import { config } from "dotenv";
 import path from "path";
-import { Fleet } from "eris-fleet";
+import { Fleet, Options } from "eris-fleet";
+import fs from "fs";
 import Eris from "eris";
+import schedule from "node-schedule";
 import _logger from "./logger";
 import { clearRestartNotification, initializeBotStatsPoster, registerProcessEvents, startWebServer } from "./helpers/management_utils";
+import storeDailyStats from "./scripts/store-daily-stats";
+import dbContext from "./database_context";
+import { reloadFactCache } from "./fact_generator";
+import { EnvType } from "./types";
+import { seedAndDownloadNewSongs } from "./seed/seed_db";
 
 const logger = _logger("index");
 
 config({ path: path.resolve(__dirname, "../.env") });
 const ERIS_INTENTS = Eris.Constants.Intents;
-const options = {
+const options: Options = {
+    whatToLog: {
+        blacklist: ["stats_update"],
+    },
     path: path.join(__dirname, "./kmq.js"),
     token: process.env.BOT_TOKEN,
     clientOptions: {
@@ -35,16 +45,32 @@ const options = {
     },
 };
 
-(async () => {
-    const Admiral = new Fleet(options);
-    if (isMaster) {
-        Admiral.on("log", (m) => logger.info(m));
-        Admiral.on("debug", (m) => logger.debug(m));
-        Admiral.on("warn", (m) => logger.warn(m));
-        Admiral.on("error", (m) => logger.error(m));
+function registerGlobalIntervals(fleet: Fleet) {
+    // everyday at 12am UTC => 7pm EST
+    schedule.scheduleJob("0 0 * * *", async () => {
+        storeDailyStats(fleet.stats?.guilds);
+        reloadFactCache();
+    });
 
-        // Logs stats when they arrive
-        Admiral.on("stats", (m) => logger.log(m));
+    // every hour
+    schedule.scheduleJob("15 * * * *", async () => {
+        if (process.env.NODE_ENV !== EnvType.PROD) return;
+        logger.info("Performing regularly scheduled Daisuki database seed");
+        const overrideFileExists = fs.existsSync(path.join(__dirname, "../../data/skip_seed"));
+        if (overrideFileExists) {
+            return;
+        }
+        await seedAndDownloadNewSongs(dbContext);
+    });
+}
+
+(async () => {
+    const fleet = new Fleet(options);
+    if (isMaster) {
+        fleet.on("log", (m) => logger.info(m));
+        fleet.on("debug", (m) => logger.debug(m));
+        fleet.on("warn", (m) => logger.warn(m));
+        fleet.on("error", (m) => logger.error(m));
 
         logger.info("Starting web servers...");
         await startWebServer();
@@ -57,5 +83,8 @@ const options = {
 
         logger.info("Clearing existing restart notifications...");
         await clearRestartNotification();
+
+        logger.info("Registering global intervals");
+        registerGlobalIntervals(fleet);
     }
 })();
